added user registration

+++ base/views.py
@@ -1,6 +1,7 @@
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth.models import User
 from django.db.models import Q
 from django.http import HttpResponse
@@ -19,8 +20,12 @@ from .models import Room, Topic
 # ]
 
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect ('home')
+
     if request.method == 'POST':
-            username = request.POST.get('username')
+            username = request.POST.get('username').lower()
             password = request.POST.get('password')
 
             try:
@@ -40,13 +45,29 @@ def loginPage(request):
                 messages.error(request, 'Invalid username or password')
 
 
-    context = {}
+    context = {'page':page}
     return render(request, 'base/login_register.html',context)
 
 
 def LogoutUser(request):
     logout(request)
     return redirect('home')
+
+
+def registerUser(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'error has occured while registering ')
+    return render(request, 'base/login_register.html',{'form':form})
 
     
 
@@ -82,10 +103,12 @@ def createRoom(request):
 
 
 
-
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    if request.user != room.host:
+        return HttpResponse('Youre not allowed here  ')
     if request.method =='POST':
         form  = RoomForm(request.POST, instance=room)
         if form.is_valid:
@@ -94,9 +117,11 @@ def updateRoom(request,pk):
     context = {'form':form}
     return render (request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('Youre not allowed here  ')
     if request.method == 'POST':
         room.delete()
         return redirect('home')
